# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Accounting Platform API
  version: 0.1.0
  description: >
    REST API for the AbhashTech multi-tenant accounting and bookkeeping
    platform with double-entry ledger, invoicing, expenses, payroll, tax,
    reporting, imports, backups, and investment accounting.
  contact:
    name: AbhashTech Accounting
    url: https://accounting.abhashtech.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.accounting.abhashtech.com/api/v1
    description: Production placeholder
...
security:
  - bearerAuth: []
paths:
...
  /organizations/{organizationId}/payroll/india-preview:
    post:
      tags: [Payroll]
      summary: Preview India payroll earnings and statutory deductions
      description: >
        Calculates a configurable India payroll component breakdown for PF, ESI,
        professional tax, and TDS. Defaults are starter values only; admins
        should override rates, wage ceilings, and fixed deductions to match the
        organization's active payroll policy before creating payroll runs.
      operationId: previewIndiaPayroll
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreviewIndiaPayrollRequest'
      responses:
        '200':
          description: India payroll component preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndiaPayrollPreview'
        default:
          $ref: '#/components/responses/Error'
...
  /organizations/{organizationId}/seed/india-defaults:
    post:
      tags: [Tax, Accounts, System]
      summary: Seed India default chart of accounts and GST presets
      operationId: seedIndiaDefaults
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: Seed result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndiaSeedResult'
        default:
          $ref: '#/components/responses/Error'
...
  /organizations/{organizationId}/users:
    get:
      tags: [Organizations]
      summary: List organization users
      operationId: listOrganizationUsers
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: Organization users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationUser'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Organizations]
      summary: Create an organization user and membership
      operationId: createOrganizationUser
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationUserRequest'
      responses:
        '201':
          description: Created organization user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationUser'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: validation_error
            message:
              type: string
              example: Request validation failed
            details:
              type: array
              items:
                type: object
                additionalProperties: true
...
    PreviewIndiaPayrollRequest:
      type: object
      properties:
        basic_minor:
          type: integer
          format: int64
          description: Basic pay in minor currency units.
          example: 2000000
        hra_minor:
          type: integer
          format: int64
          description: House rent allowance in minor currency units.
          example: 500000
        special_minor:
          type: integer
          format: int64
          description: Special allowance in minor currency units.
          example: 250000
        bonus_minor:
          type: integer
          format: int64
          example: 0
        reimbursement_minor:
          type: integer
          format: int64
          example: 0
        employee_pf_enabled:
          type: boolean
          default: true
        employee_pf_rate_bps:
          type: integer
          format: int64
          description: Employee PF rate in basis points. Zero uses the API default starter value.
          example: 1200
        pf_wage_ceiling_minor:
          type: integer
          format: int64
          description: PF wage ceiling in minor currency units. Zero uses the API default starter value.
          example: 1500000
        employee_esi_enabled:
          type: boolean
          default: true
        employee_esi_rate_bps:
          type: integer
          format: int64
          description: Employee ESI rate in basis points. Zero uses the API default starter value.
          example: 75
        esi_gross_limit_minor:
          type: integer
          format: int64
          description: ESI gross wage eligibility ceiling in minor currency units. Zero uses the API default starter value.
          example: 2100000
        professional_tax_minor:
          type: integer
          format: int64
          description: Fixed professional tax deduction in minor currency units.
          example: 20000
        tds_minor:
          type: integer
          format: int64
          description: Fixed TDS deduction in minor currency units.
          example: 150000
    IndiaPayrollPreview:
      type: object
      required: [gross_pay_minor, deductions_minor, net_pay_minor, components, rule_summary]
      properties:
        gross_pay_minor:
          type: integer
          format: int64
        deductions_minor:
          type: integer
          format: int64
        net_pay_minor:
          type: integer
          format: int64
        components:
          type: array
          items:
            $ref: '#/components/schemas/CreatePayrollComponentRequest'
        rule_summary:
          $ref: '#/components/schemas/IndiaPayrollRuleSummary'
    IndiaPayrollRuleSummary:
      type: object
      required: [employee_pf_enabled, employee_pf_rate_bps, pf_wage_ceiling_minor, employee_esi_enabled, employee_esi_rate_bps, esi_gross_limit_minor, professional_tax_minor, tds_minor]
      properties:
        employee_pf_enabled:
          type: boolean
        employee_pf_rate_bps:
          type: integer
          format: int64
        pf_wage_ceiling_minor:
          type: integer
          format: int64
        employee_esi_enabled:
          type: boolean
        employee_esi_rate_bps:
          type: integer
          format: int64
        esi_gross_limit_minor:
          type: integer
          format: int64
        professional_tax_minor:
          type: integer
          format: int64
        tds_minor:
          type: integer
          format: int64
...
    CreatePayrollComponentRequest:
      type: object
      required: [code, name, type, amount_minor]
      properties:
        code:
          type: string
          example: BASIC
        name:
          type: string
          example: Basic Pay
        type:
          type: string
          enum: [earning, deduction]
        amount_minor:
          type: integer
          format: int64
        is_statutory:
          type: boolean
          default: false
...
    IndiaSeedResult:
      type: object
      required: [accounts_created, tax_rates_created, tax_groups_created, tax_authority_created]
      properties:
        accounts_created:
          type: integer
        tax_rates_created:
          type: integer
        tax_groups_created:
          type: integer
        tax_authority_created:
          type: boolean
...
    OrganizationUser:
      type: object
      required: [user_id, organization_id, name, email, role, is_active]
      properties:
        user_id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, accountant, bookkeeper, payroll_manager, viewer, employee_self_service]
        is_active:
          type: boolean
    CreateOrganizationUserRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 12
        role:
          type: string
          enum: [admin, accountant, bookkeeper, payroll_manager, viewer, employee_self_service]
...